refactor(playground): remove dead chat branch from agent overview page
The overview route is the only consumer of the agent index page now that
chat lives on the standalone thread page, so the chat rendering path,
AgentChatShell, view props, thread state, and the chat/settings view
transition CSS are no longer reachable. Memory sidebar styles move to
their own file.
