+++ packages/playground/src/domains/agents/components/memory-sidebar/memory-sidebar.css
@@ -1,0 +1,52 @@
+/* Memory sidebar overlay: dialog-like expansion without shifting the threads layer */
+.memory-sidebar-overlay {
+  transition-property: height, margin, border-radius, border-color, background-color, box-shadow;
+  transition-duration: 420ms;
+  transition-timing-function: cubic-bezier(0.16, 1, 0.3, 1);
+  will-change: height, margin, border-radius;
+}
+
+.memory-sidebar-thread-layer {
+  transition: opacity 220ms cubic-bezier(0.22, 1, 0.36, 1);
+}
+
+.memory-sidebar-overlay:not(:has(.memory-card-content)) {
+  animation: memory-sidebar-hover-lock 420ms steps(1, end);
+}
+
+@keyframes memory-sidebar-hover-lock {
+  from,
+  to {
+    background-color: var(--surface4);
+  }
+}
+
+/* Memory card expansion: the detail content fades in once the card has grown */
+@keyframes memory-card-content-enter {
+  from {
+    opacity: 0;
+  }
+
+  to {
+    opacity: 1;
+  }
+}
+
+.memory-card-content {
+  animation: memory-card-content-enter 320ms cubic-bezier(0.16, 1, 0.3, 1) 80ms both;
+}
+
+@media (prefers-reduced-motion: reduce) {
+  .memory-sidebar-overlay,
+  .memory-sidebar-thread-layer {
+    transition: none;
+  }
+
+  .memory-sidebar-overlay:not(:has(.memory-card-content)) {
+    animation: none;
+  }
+
+  .memory-card-content {
+    animation: none;
+  }
+}
